+++ CHANGELOG.rst
@@ -5,6 +5,6 @@ Changelog
 Version 0.1
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- initiate SiModIn!
+- add interface
+- add DWSIM and Tespy examples